Transform GameManager in ConnectionsManager
- Transform GameManager in ConnectionsManager
- Move BlockingQueue in its own file
- Some code improvements
